Move common parts to module_utils
As we need `parse_config` also for further lookup functions we moved it
to a `modules_utils` file.
